+++ pcgl2cache/__init__.py
@@ -1,3 +1,3 @@
 __version__ = "0.0.1"
 
-from .proc import attributes
+from .core import attributes
