+++ pcgl2cache/ingest/v1/jobs.py
@@ -50,7 +50,7 @@ def enqueue_atomic_tasks(imanager: IngestionManager, cv_path: str):
 
 
 def _ingest_chunk(im_info: str, cv_path: str, chunk_coord: Sequence[int]):
-    from ...proc.calc_l2_feats import run_l2cache
+    from ...core.calc_l2_feats import run_l2cache
 
     imanager = IngestionManager.from_pickle(im_info)
     chunk_coord = np.array(list(chunk_coord), dtype=np.int)
@@ -59,8 +59,8 @@ def _ingest_chunk(im_info: str, cv_path: str, chunk_coord: Sequence[int]):
 
 
 def _ingest_chunks(im_info: str, cv_path: str, chunk_coords: Sequence[Sequence[int]]):
-    from ...proc.calc_l2_feats import run_l2cache_batch
-    from ...proc.calc_l2_feats import write_to_db
+    from ...core.calc_l2_feats import run_l2cache_batch
+    from ...core.calc_l2_feats import write_to_db
     from kvdbclient import BigTableClient
 
     imanager = IngestionManager.from_pickle(im_info)
